🐛 fix : 페이지 전환 시 타이머 사라지는 문제 해결 (useEffect 코드 이동) 및 JSDoc 상세 추가
